feat(ai): promote NVIDIA InferenceHub to a built-in provider
Move NVIDIA InferenceHub from a coding-agent extension example to a
first-class provider in @earendil-works/pi-ai so every pi consumer
(coding-agent, RPC/HTTP embeddings, third-party pi-ai users) can use it.
The provider registers with an empty baseline and uses createProvider's
`fetchModels` hook to pull the model catalog from the InferenceHub
metadata API on refresh, routing each model to its own LiteLLM proxy URL
under either `openai-completions` or `openai-responses`.

Fixes two coding-agent bugs uncovered along the way:

- `withRemoteCatalog` unconditionally wrapped every built-in provider
  except `radius`, silently replacing any provider-owned `refreshModels`
  with the pi.dev catalog fetcher. Switch the guard to skip providers
  that already carry a `refreshModels` hook so future dynamic providers
  do not need to be added to a hardcoded allowlist.
- `--list-models` returned the current in-memory snapshot without ever
  triggering a network refresh, so dynamic catalogs stayed empty. Force
  a network refresh before rendering the list.

Env vars (both from InferenceHub Portal's key-management page):
  NVIDIA_INFERENCEHUB_API_KEY       — required, Bearer token for LLM calls
  NVIDIA_INFERENCEHUB_METADATA_KEY  — optional, X-API-Key for catalog
