+++ packages/coding-agent/examples/extensions/README.md
@@ -129,7 +129,6 @@ cp permission-gate.ts ~/.pi/agent/extensions/
 |-----------|-------------|
 | `custom-provider-anthropic/` | Custom Anthropic provider with OAuth support and custom streaming implementation |
 | `custom-provider-gitlab-duo/` | GitLab Duo provider using pi-ai's built-in Anthropic/OpenAI streaming via proxy |
-| `nvidia-inference-hub.ts` | NVIDIA InferenceHub provider — fetches the model catalog and routes each model to its own LiteLLM proxy URL |
 
 ### External Dependencies
